% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n2_theta_stop.R
\name{get_n2_theta_stop}
\alias{get_n2_theta_stop}
\title{Compute the n2 and Theta_stop values}
\usage{
get_n2_theta_stop(HR_eff = 0.55, HR_ineff = 0.9, PoS_go = 0.9, PoS_nogo = 0.85)
}
\arguments{
\item{HR_eff}{HR threshold for efficacy}

\item{HR_ineff}{HR threshold for inefficacy}

\item{PoS_go}{Probability of Success for Go decision}

\item{PoS_nogo}{Probability of Success for NoGo decision}
}
\value{
A list with the values Theta_eff, Theta_ineff,
Go, NoGo, n2, Theta_stop
}
\description{
Compute the n2 and Theta_stop values
}
