% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos.R
\name{get_pos}
\alias{get_pos}
\title{Compute the probability of Success?}
\usage{
get_pos(Theta_stop = 0.32, Theta_eff = 0.59, n2 = 89, n3 = 90, alpha = 0.05)
}
\arguments{
\item{Theta_stop}{Theta_stop value}

\item{Theta_eff}{Theta_eff value}

\item{n2}{n2 value}

\item{n3}{n2 value}

\item{alpha}{alpha value}
}
\value{
A scalar value that is the probability of
Success?
}
\description{
Obviously I would need to read the paper again
}
